Feat: Add initial project structure and settings
Spring Boot (Java 21)
- 6 core dependencies (Web, JPA, Security, Validation, etc.)
- 7 core packages (config, entity, controller, etc.)
- MySQL DB connection in application.yml
- Standard .gitignore file
API 4-3 끝
보안설정 끝

+++ src/main/java/teamprojects/demo/repository/StudyApplicationRepository.java
@@ -9,23 +9,23 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
+// ⭐️ <StudyApplication, Integer> 로 선언되어 있음 (ID 타입 = Integer)
 public interface StudyApplicationRepository extends JpaRepository<StudyApplication, Integer> {
 
-    // ⭐️ 1. 이 메서드를 추가해야 UserService.login이 정상 작동합니다.
-    // ⭐️ (로그인 시, 스터디장(Leader)에게 온 'PENDING' 상태의 신청 건수를 조회)
+    // ⭐️ [완벽함] 로그인 시 알림 개수 조회 쿼리
     @Query("SELECT COUNT(sa) FROM StudyApplication sa JOIN sa.study s WHERE s.leader.id = :leaderId AND sa.status = 'PENDING'")
     Integer countPendingApplicationsByLeaderId(@Param("leaderId") Integer leaderId);
 
-    // API 2-2 User, Study, Status(String)로 신청 내역 존재 여부 확인
-    // (주의: Status가 Enum이 아니라 String으로 설계되었다고 가정했습니다. Enum이면 타입 변경 필요)
+    // API 2-2 신청 내역 존재 여부
     boolean existsByUserAndStudyAndStatus(User user, Study study, String status);
 
-    // (API 3-1: 신청자 목록 조회 시 사용)
+    // API 3-1 신청자 목록 조회
     List<StudyApplication> findByStudyAndStatus(Study study, String status);
 
-    // (API 2-3: 중복 신청 방지 시 사용)
+    // API 2-3 중복 신청 방지
     boolean existsByUserAndStudy(User user, Study study);
 
-    // (API 3-2, 3-3: 승인/거절 시 특정 신청서 조회)
-    Optional<StudyApplication> findByIdAndStudy(Long applicationId, Study study);
+    // 🚨 [수정 완료] Long -> Integer로 변경해야 합니다!
+    // (위에서 JpaRepository<..., Integer>로 선언했기 때문입니다.)
+    Optional<StudyApplication> findByIdAndStudy(Integer applicationId, Study study);
 }
